refactor: Modify Package architecture to MVC
