+++ src/main/java/racingcar/domain/RandomEngine.java
@@ -1,5 +1,7 @@
 package racingcar.domain;
 
+import racingcar.util.RandomGenerator;
+
 public class RandomEngine implements Engine{
     @Override
     public int getPower() {
