+++ src/test/java/racingcar/RandomGeneratorTest.java
@@ -2,7 +2,7 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.domain.RandomGenerator;
+import racingcar.util.RandomGenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
